REFACT: Item 일부 구조 개선
